+++ CHANGELOG.md
@@ -11,6 +11,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - On a typical system during daylight hours, the Envoy will execute the task within 15 minutes.
 - More info about Power Production task here: https://support.enphase.com/s/article/How-do-I-disable-and-enable-power-production
 
+## [7.0.0] - (14.07.2023)
+## After Update to this version need to make corespondent changes in plugin config!!!
+## Changes
+- added support to get JWT Token automatically from enlighten server using user credentials data
+- added support to check expired JWT Token and get new if expired
+- added debug for RESTFul server
+- added `token` to the RESTFul server request
+- added `Token` to the MQTT publisher
+- config schema updated
+- cleanup
+### Removed properties
+- `envoyFirmware7xxToken`
+### Added properties
+- `enlightenUser`
+- `enlightenPasswd`
+- `envoySerialNumber`
+
 ## [6.7.0] - (27.02.2023)
 ## Changes
 - added powerful RESTFul server to use with own automations